--- src/main/java/org/promgrammers/springbootbasic/domain/Customer.java
@@ -1,7 +1,0 @@
-package org.promgrammers.springbootbasic.domain;
-
-import java.util.UUID;
-
-public record Customer(UUID customerId, CustomerType customerType) {
-
-}
