--- src/main/java/org/promgrammers/springbootbasic/domain/CustomerType.java
@@ -1,5 +1,0 @@
-package org.promgrammers.springbootbasic.domain;
-
-public enum CustomerType {
-    WHITE, BLACK
-}
