+++ src/main/java/org/promgrammers/springbootbasic/domain/customer/model/Customer.java
@@ -1,0 +1,7 @@
+package org.promgrammers.springbootbasic.domain.customer.model;
+
+import java.util.UUID;
+
+public record Customer(UUID customerId, CustomerType customerType) {
+
+}
