+++ src/main/java/org/promgrammers/springbootbasic/domain/customer/model/CustomerType.java
@@ -1,0 +1,5 @@
+package org.promgrammers.springbootbasic.domain.customer.model;
+
+public enum CustomerType {
+    WHITE, BLACK
+}
