+++ src/main/java/org/promgrammers/springbootbasic/domain/voucher/repository/VoucherRepository.java
@@ -1,0 +1,17 @@
+package org.promgrammers.springbootbasic.domain.voucher.repository;
+
+import org.promgrammers.springbootbasic.domain.voucher.model.Voucher;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface VoucherRepository {
+    Optional<Voucher> findById(UUID voucherId);
+
+    Voucher insert(Voucher voucher);
+
+    List<Voucher> findAll();
+
+    void deleteAll();
+}
