+++ src/test/java/org/promgrammers/springbootbasic/controller/CommandProgramStatusTest.java
@@ -7,6 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class CommandProgramStatusTest {
+
     @Test
     @DisplayName("isStop 메서드 호출시 isRunning이 false가 된다")
     void isStopTest() throws Exception {
