+++ src/test/java/org/promgrammers/springbootbasic/controller/CommandTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.promgrammers.springbootbasic.view.Command;
+import org.promgrammers.springbootbasic.domain.voucher.model.Command;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
